+++ src/Providers/UiServiceProvider.php
@@ -18,6 +18,7 @@ use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
 use Konekt\AppShell\EnumIcons;
 use Konekt\AppShell\Icons\FontAwesome6IconTheme;
+use Konekt\AppShell\Icons\FontAwesome6ProIconTheme;
 use Konekt\AppShell\Icons\FontAwesomeIconTheme;
 use Konekt\AppShell\Icons\LineIconsTheme;
 use Konekt\AppShell\Icons\TablerIconTheme;
@@ -43,6 +44,7 @@ class UiServiceProvider extends ServiceProvider
         IconThemes::add(ZmdiIconTheme::ID, ZmdiIconTheme::class);
         IconThemes::add(FontAwesomeIconTheme::ID, FontAwesomeIconTheme::class);
         IconThemes::add(FontAwesome6IconTheme::ID, FontAwesome6IconTheme::class);
+        IconThemes::add(FontAwesome6ProIconTheme::ID, FontAwesome6ProIconTheme::class);
         IconThemes::add(LineIconsTheme::ID, LineIconsTheme::class);
         IconThemes::add(TablerIconTheme::ID, TablerIconTheme::class);